+++ sunsynk/sensors.py
@@ -77,12 +77,15 @@ class MathSensor(Sensor):
 
     factors: Tuple[float, ...] = attr.field(default=None, converter=ensure_tuple)
     no_negative: bool = attr.field(default=False)
+    absolute: bool = attr.field(default=False)
 
     def update_value(self) -> None:
         """Update the value."""
         self.value = int_round(
             sum(signed(i) * s for i, s in zip(self.reg_value, self.factors))
         )
+        if self.absolute and not isinstance(self.value, str) and self.value < 0:
+            self.value = -self.value
         if self.no_negative and not isinstance(self.value, str) and self.value < 0:
             self.value = 0